Update MainFrame.java
This is my updated version of the MainFrame. As one can see, I removed a lot of aspects of the first MainFrame draft as it was buggy and cluttered. Although it does not do much now, the comments provide a sufficient explanation of what is occurring. The reason I made this edit into a new branch is so that Lucca can have his version and mine side by side in order to add what ever I removed as efficiently as possible. Thanks Lucca!