IMPORTANT Update GameFrame 11-15-22
This is the most recent GameFrame update with an actual start to the game. The deck is shuffled and 4 tiles are handed out to each factory successfully, randomly each time the game is rebooted. All of the images for all of the tiles in a factory are accurately represented on screen as well.